go.mod: add dependencies for go tooling
This is a change backported from #5792.

Under bzlmod, there will be no `deps.bzl` file.
The `go_repository` repos are automatically generated with gazelle's
`go_deps` extension parsing `go.mod` and `go.sum` files.
This let us consolidate go dependency management to a single source of
truth that is consistent with our editor experience.

The downside of this approach is that we should not be adding
`go_repository` declarations manually, but instead, adding them to
`go.mod` file by calling `go mod tidy -e`.
Create a dummy Go source file that imports all the Go tools dependencies
and guard it with a Go build directive that will not be used.