ci_runner/bundle: stop using runfiles
In the past, we used to maintain a separate local Executor build
where `buildbuddy_ci_runner` binary is accessed via runfiles. However,
in production, we ship our Executor as a standalone binary with and
embedded FS inside the binary.

With the upcoming changes to how runfiles are accessed in bzlmod, it
would be simpler for us to retire the local build with runfiles in this
code path.

Make sure that we always use the embedsrcs to access
`buildbuddy_ci_runner` binary.